Confirmation messages for admin commands are now moved into the convenience methods in NotificationManager
