+++ src/main/resources/locale/locale_en_US.properties
@@ -829,11 +829,12 @@ Commands.xprate.started.1=[[GOLD]]mcMMO XP RATE IS NOW {0}x!
 
 # Admin Notifications
 Server.ConsoleName=[Server Console]
-Notifications.Admin.XPRate.Start.Self=[[GREEN]]You have set the XP rate to [[GOLD]]{0}[[GREEN]]x!
+Notifications.Admin.XPRate.Start.Self=[[GRAY]]You have set the XP rate to [[GOLD]]{0}[[GREEN]]x!
 Notifications.Admin.XPRate.End.Self=[[GRAY]]You ended the XP rate event.
 Notifications.Admin.XPRate.End.Others=The user {0} [[GRAY]]has ended the XP rate event
 Notifications.Admin.XPRate.Start.Others=The user {0} [[GRAY]]has started or modified an XP rate event with properties: {1}x
-Notifications.Admin.Format=[[GOLD]]([[GREEN]]mcMMO [[DARK_AQUA]]Admin Notification[[GOLD]]) [[GRAY]]{0}
+Notifications.Admin.Format.Others=[[GOLD]]([[GREEN]]mcMMO [[DARK_AQUA]]Admin Notification[[GOLD]]) [[GRAY]]{0}
+Notifications.Admin.Format.Self=[[GOLD]]([[GREEN]]mcMMO [[DARK_RED]]Admin Command Executed[[GOLD]]) [[GRAY]]{0}
 
 # Event
 XPRate.Event=[[GOLD]]mcMMO is currently in an XP rate event! XP rate is {0}x!
